Memory management.

+++ idt.h
@@ -15,16 +15,18 @@
   Justin R. Wilson
 */
 
+#include "types.h"
+
 /* Operating system trap starts at 128. */
 #define TRAP_BASE 128
 
 struct registers
 {
-  unsigned int ds;
-  unsigned int edi, esi, ebp, esp, ebx, edx, ecx, eax;
-  unsigned int number;
-  unsigned int error;
-  unsigned int eip, cs, eflags, useresp, ss;
+  uint32_t ds;
+  uint32_t edi, esi, ebp, esp, ebx, edx, ecx, eax;
+  uint32_t number;
+  uint32_t error;
+  uint32_t eip, cs, eflags, useresp, ss;
 };
 typedef struct registers registers_t;
 typedef void (*interrupt_handler_t) (registers_t*);
@@ -39,7 +41,7 @@ void
 disable_interrupts (void);
 
 void
-set_interrupt_handler (unsigned int num,
+set_interrupt_handler (uint8_t num,
 		       interrupt_handler_t handler);
 
 #endif /* __interrupt_h__ */
